# packages/contracts/generated/openapi.v1.json
{
  "components": {
...
    "schemas": {
...
      "DoctorCoverage": {
        "properties": {
          "coveredDoctorProfileId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "coveringDoctorProfileId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "endsAt": {
            "example": "2026-09-18T10:49:42.206Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "rowVersion": {
            "type": "integer"
          },
          "startsAt": {
            "example": "2026-09-18T10:49:42.206Z",
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "REVOKED"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "coveredDoctorProfileId",
          "coveringDoctorProfileId",
          "startsAt",
          "endsAt",
          "reason",
          "status",
          "rowVersion"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "code",
          "message"
        ],
        "type": "object"
      },
      "GrantCoverageRequest": {
        "properties": {
          "coveredDoctorProfileId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "coveringDoctorProfileId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string"
          },
          "reason": {
            "maxLength": 300,
            "minLength": 1,
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "coveredDoctorProfileId",
          "coveringDoctorProfileId",
          "startsAt",
          "endsAt",
          "reason"
        ],
        "type": "object"
      },
...
      "Membership": {
        "properties": {
          "chamberIds": {
            "items": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "clinicIds": {
            "items": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/MembershipPermissionOverrides"
          },
          "role": {
            "enum": [
              "tenant_owner",
              "clinic_admin",
              "doctor",
              "nurse",
              "receptionist",
              "billing_manager"
            ],
            "type": "string"
          },
          "rowVersion": {
            "type": "integer"
          },
          "status": {
            "enum": [
              "INVITED",
              "ACTIVE",
              "SUSPENDED",
              "REMOVED"
            ],
            "type": "string"
          },
          "userId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "displayName",
          "email",
          "role",
          "status",
          "permissions",
          "clinicIds",
          "chamberIds",
          "rowVersion"
        ],
        "type": "object"
      },
      "MembershipPermissionOverrides": {
        "properties": {
          "denials": {
            "items": {
              "maxLength": 64,
              "type": "string"
            },
            "maxItems": 80,
            "type": "array"
          },
          "grants": {
            "items": {
              "maxLength": 64,
              "type": "string"
            },
            "maxItems": 80,
            "type": "array"
          }
        },
        "required": [
          "grants",
          "denials"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "Error body for every non-2xx response (API-IMPLEMENTATION §1)",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object"
          },
          "fieldErrors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "description": "Safe, localizable message key (never PHI)",
            "type": "string"
          },
          "requestId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "retryAfterSeconds": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "ResponseMeta": {
        "properties": {
          "replayed": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "requestId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
      "RevokeCoverageRequest": {
        "properties": {
          "expectedRowVersion": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "expectedRowVersion"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Access token (EdDSA JWT). Web keeps it in memory; mobile in secure storage.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HMedic clinic platform API. Generated from packages/contracts (do not edit the JSON by hand). Business routes live under /api/v1; health routes at the root.",
    "title": "HMedic API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/doctor-coverages": {
      "get": {
        "operationId": "listCoverages",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/DoctorCoverage"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Coverages"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "grantCoverage",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCoverageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DoctorCoverage"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Granted"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/doctor-coverages/{id}/revoke": {
      "post": {
        "operationId": "revokeCoverage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeCoverageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DoctorCoverage"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Revoked"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
...
    "/api/v1/memberships": {
      "get": {
        "operationId": "listMemberships",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Membership"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Memberships (membership.manage)"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "createMembership",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Membership"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/memberships/{id}": {
      "patch": {
        "operationId": "updateMembership",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Membership"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development",
      "url": "http://localhost:3000"
    }
  ],
...
}